+++ src/app/components/SettingsMenu/index.tsx
@@ -117,7 +117,7 @@ const listStyle: React.CSSProperties = {
     overflowY: 'visible',
 };
 
-export const C_VERSION = '0.30.13';
+export const C_VERSION = '0.30.14';
 export const C_CUSTOM_BG_ID = 'river_custom_bg';
 export const C_AVATAR_SIZE = 640;